TAB only works in prefix and substring searches
Also, rename some variables for consistency and ensure that the search
buffer is nul terminated.

+++ fgetc.c
@@ -20,7 +20,7 @@ enum
 
 static unsigned char getc_buffer[GETC_BUFF_SIZE] = { '\0' };
 
-static long next_buffer_pos = 0; /* Next free position in the getc buffer. */
+static int next_buffer_pos = 0; /* Next free position in the getc buffer. */
 
 /* ======================================== */
 /* Gets a (possibly pushed-back) character. */
